requirements.txt
numpy==1.18.4
tqdm==4.46.0
opencv_contrib_python==4.5.1.48
pandas==1.0.3
matplotlib==3.1.3
seaborn==0.10.1
tensorflow_gpu==2.2.0
Keras==2.4.3
scikit_learn==0.24.1